Started Blackjack detection.

+++ src/nbradham/blackjack/Player.java
@@ -4,7 +4,7 @@ import nbradham.blackjack.Game.Action;
 
 sealed interface Player permits TerminalPlayer {
 
-	byte getBet();
+	int getBet();
 
 	void setGame(Game game);
